chore(services): rename Gemini service to LiteLLM service
