+++ src/cogs/auto_translation.py
@@ -1,13 +1,13 @@
 import re
 from discord.ext import commands
 from discord import Message
-from services.gemini_service import GeminiService, LLMFallbackError
+from services.litellm_service import LiteLLMService, LLMFallbackError
 
 
 class AutoTranslationCog(commands.Cog, name="ArabicTranslate"):
     def __init__(self, bot: commands.Bot):
         self.bot = bot
-        self.gemini_service = GeminiService()
+        self.llm_service = LiteLLMService()
         self.arabic_pattern = re.compile(r"[\u0600-\u06FF]")
 
     def _get_translation_args(self, text: str):
@@ -22,7 +22,7 @@ class AutoTranslationCog(commands.Cog, name="ArabicTranslate"):
     def _translate_with_gemini(self, text: str) -> str | None:
         """Send Arabic text to Gemini and return English translation."""
         history, instr = self._get_translation_args(text)
-        translation = self.gemini_service.make_gemini_request(history, instr)
+        translation = self.llm_service.make_gemini_request(history, instr)
 
         if not translation or translation.startswith("Sorry,"):
             print(f"ArabicTranslateFeature: Translation error: {translation}")
@@ -32,7 +32,7 @@ class AutoTranslationCog(commands.Cog, name="ArabicTranslate"):
     def _translate_with_ollama(self, text: str) -> str | None:
         """Fallback translation method using local Ollama model."""
         history, instr = self._get_translation_args(text)
-        translation = self.gemini_service.make_ollama_request(history, instr)
+        translation = self.llm_service.make_ollama_request(history, instr)
 
         if not translation or translation.startswith("Sorry,"):
             print(f"ArabicTranslateFeature: Ollama translation error: {translation}")
